added folders, images, globalstyle and textdata.
