+++ src/components/pages/About/About.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import MetaDecorator from '../../MetaDecorator/MetaDecorator';
+
+const About = () => {
+    return (
+        <div>
+            <MetaDecorator
+             description='Here is About Page'
+             title='My About Page'
+             />
+           
+            <p>Hola! About here</p>
+        </div>
+    )
+}
+
+export default About;
