+++ src/components/pages/Coaches/Coaches.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import MetaDecorator from '../../MetaDecorator/MetaDecorator';
+
+const Coaches = () => {
+    return (
+        <div>
+            <MetaDecorator
+            description='Here is Coaches Page'
+            title='My Coaches Page'
+            />
+
+          <p>Hola! Coaches here</p>
+        </div>
+    )
+}
+
+export default Coaches;
