+++ src/data/textData.js
@@ -1,0 +1,1 @@
+export const textData = {}
